0809 공부내용

+++ UnitTest/Character/Knight.h
@@ -22,7 +22,7 @@ private:
 	float Fspeed;
 
 	float delay;
-	double angle_attack;
+	float angle_attack;
 
 
 };
